+++ app/src/main/AndroidManifest.xml
@@ -17,13 +17,13 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".MapErrorActivity"
+            android:name=".presentation.activity.MapErrorActivity"
             android:exported="false" />
         <activity
-            android:name=".SearchActivity"
+            android:name=".presentation.activity.SearchActivity"
             android:exported="false" />
         <activity
-            android:name=".KakaoMapView"
+            android:name=".presentation.activity.KakaoMapViewActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
