+++ app/src/main/java/campus/tech/kakao/map/data/RetrofitService.kt
@@ -1,5 +1,6 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.data
 
+import campus.tech.kakao.map.domain.model.KakaoData
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Header
