+++ app/src/main/java/campus/tech/kakao/map/data/SearchDbHelper.kt
@@ -1,16 +1,13 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.data
 
 import android.content.ContentValues
 import android.content.Context
 import android.database.sqlite.SQLiteDatabase
 import android.database.sqlite.SQLiteOpenHelper
 import android.util.Log
-import android.view.View
+import campus.tech.kakao.map.domain.model.SearchData
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
 import retrofit2.Retrofit
 import retrofit2.awaitResponse
 import retrofit2.converter.gson.GsonConverterFactory
