+++ app/src/main/java/campus/tech/kakao/map/presentation/adapter/SavedSearchAdapter.kt
@@ -1,12 +1,12 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.presentation.adapter
 
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.AdapterView
 import android.widget.Button
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import campus.tech.kakao.map.R
 
 class SavedSearchAdapter() : RecyclerView.Adapter<SavedSearchAdapter.ViewHolder>() {
