+++ app/src/main/java/campus/tech/kakao/map/presentation/viewmodel/MainViewModel.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.presentation.viewmodel
 
 import android.content.ContentValues
 import androidx.lifecycle.ViewModel
