+++ app/src/main/res/layout/activity_kakao_map_view.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".KakaoMapView">
+    tools:context=".presentation.activity.KakaoMapViewActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
